feat: refactor validation into the validation service

+++ api/src/main/java/demo/service/UserService.java
@@ -1,14 +1,10 @@
 package demo.service;
 
-import demo.exception.BusinessException;
 import demo.exception.NotFoundException;
 import demo.model.Sector;
 import demo.model.User;
 import demo.repository.UserRepository;
-import demo.service.validation.ValidationResult;
-import demo.service.validation.sector_validator.SectorValidator;
-import demo.service.validation.user_validator.UserIdValidator;
-import demo.service.validation.user_validator.UserValidator;
+import demo.service.validation.ValidationService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -23,45 +19,11 @@ import java.util.List;
 @RequiredArgsConstructor
 public class UserService {
     private final UserRepository userRepository;
-    private final List<UserValidator> userValidators;
-    private final List<UserIdValidator> userIdValidators;
-
-    protected void validateUserData(User user) {
-        ValidationResult validationResult = userValidators.stream()
-                .map(userValidator -> userValidator.validate(user))
-                .filter(result -> !result.isValid())
-                .findFirst()
-                .orElse(new ValidationResult().setValid(true)); // If no validation failure, return a successful result
-
-        validationCleanup(validationResult);
-    }
-
-    protected void validateUserId(Long id) {
-        ValidationResult validationResult = userIdValidators.stream()
-                .map(userValidator -> userValidator.validate(id))
-                .filter(result -> !result.isValid())
-                .findFirst()
-                .orElse(new ValidationResult().setValid(true)); // If no validation failure, return a successful result
-
-        validationCleanup(validationResult);
-    }
-
-    private void validationCleanup(ValidationResult validationResult) {
-        if (!validationResult.isValid()) {
-            log.warn("user validation failed: {}", validationResult.getMessage());
-            throw new BusinessException("DEFAULT_ERROR") {
-                // Override getMessage() to provide a custom error message
-                @Override
-                public String getMessage() {
-                    return validationResult.getMessage().getKood();
-                }
-            };
-        }
-    }
+    private final ValidationService validationService;
 
     @Transactional
     public User save(User user) {
-        validateUserData(user);
+        validationService.validateInput(user, validationService.getUserValidator());
         return userRepository.save(user);
     }
 
@@ -69,15 +31,19 @@ public class UserService {
         return userRepository.findAll(pageable);
     }
 
-    @Transactional
-    public User update(User entity, Long userId) {
-        User existingUser = userRepository.findById(userId)
+    public User findById(Long id) {
+        return userRepository.findById(id)
                 .orElseThrow(() -> {
-                    log.warn("update validation exception: given user does not exist {}", userId);
-                    return new NotFoundException("given user does not exist") {
+                    log.warn("User not found: {}", id);
+                    return new NotFoundException("User not found") {
                     };
                 });
-        validateUserData(entity);
+    }
+
+    @Transactional
+    public User update(User entity, Long userId) {
+        User existingUser = findById(userId);
+        validationService.validateInput(entity, validationService.getUserValidator());
 
         return existingUser
                 .setName(entity.getName())
@@ -87,7 +53,11 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        validateUserId(id);
+        if (!userRepository.existsById(id)) {
+            log.warn("User not found: {}", id);
+            throw new NotFoundException("User not found") {
+            };
+        }
         userRepository.deleteById(id);
     }
